Added library.json for FreematicsPlus library.
Not sure if the library.properties is correct.  Also didn't see
FreematicsPlus as being available in Platformio.  I have no experience
adding libraries to the Platformio registry so you would have to
sanity check this.

In any case I didn't want to add the library to platformio without
talking to you first, hence the pull request.